+++ src/datastructures/ParticlesBase.cpp
@@ -10,24 +10,24 @@ std::string ParticlesBase::get_structure_name () {
     return m_stucture_name;
 }
 
-void ParticlesBase::set_algorithm (const std::function<bool(int &position1x,
-                                                            int &position1y,
-                                                            int &position1z,
-                                                            int &velocity1x,
-                                                            int &velocity1y,
-                                                            int &velocity1z,
-                                                            int &acceleration1x,
-                                                            int &acceleration1y,
-                                                            int &acceleration1z,
-                                                            int &position2x,
-                                                            int &position2y,
-                                                            int &position2z,
-                                                            int &velocity2x,
-                                                            int &velocity2y,
-                                                            int &velocity2z,
-                                                            int &acceleration2x,
-                                                            int &acceleration2y,
-                                                            int &acceleration2z)> &algorithm) {
+void ParticlesBase::set_algorithm (const std::function<bool(float &position1x,
+                                                            float &position1y,
+                                                            float &position1z,
+                                                            float &velocity1x,
+                                                            float &velocity1y,
+                                                            float &velocity1z,
+                                                            float &acceleration1x,
+                                                            float &acceleration1y,
+                                                            float &acceleration1z,
+                                                            float &position2x,
+                                                            float &position2y,
+                                                            float &position2z,
+                                                            float &velocity2x,
+                                                            float &velocity2y,
+                                                            float &velocity2z,
+                                                            float &acceleration2x,
+                                                            float &acceleration2y,
+                                                            float &acceleration2z)> &algorithm) {
     m_algorithm = algorithm;
 }
 ParticlesBase::ParticlesBase (s_simulator_options *p_options, glm::vec3 *p_bounds)
