+++ src/datastructures/ParticlesBase.hpp
@@ -10,24 +10,24 @@
 
 class ParticlesBase {
     protected:
-    std::function<bool(int &position1x,
-                       int &position1y,
-                       int &position1z,
-                       int &velocity1x,
-                       int &velocity1y,
-                       int &velocity1z,
-                       int &acceleration1x,
-                       int &acceleration1y,
-                       int &acceleration1z,
-                       int &position2x,
-                       int &position2y,
-                       int &position2z,
-                       int &velocity2x,
-                       int &velocity2y,
-                       int &velocity2z,
-                       int &acceleration2x,
-                       int &acceleration2y,
-                       int &acceleration2z)>
+    std::function<bool(float &position1x,
+                       float &position1y,
+                       float &position1z,
+                       float &velocity1x,
+                       float &velocity1y,
+                       float &velocity1z,
+                       float &acceleration1x,
+                       float &acceleration1y,
+                       float &acceleration1z,
+                       float &position2x,
+                       float &position2y,
+                       float &position2z,
+                       float &velocity2x,
+                       float &velocity2y,
+                       float &velocity2z,
+                       float &acceleration2x,
+                       float &acceleration2y,
+                       float &acceleration2z)>
                          m_algorithm;
     std::string          m_stucture_name;
     s_simulator_options *m_options;
@@ -41,24 +41,24 @@ class ParticlesBase {
     virtual void add_particle (glm::vec3 p_position, glm::vec3 p_velocity, glm::vec3 p_acceleration) = 0;
     void        print_structure_name ();
     std::string get_structure_name ();
-    void set_algorithm (const std::function<bool(int &position1x,
-                                                 int &position1y,
-                                                 int &position1z,
-                                                 int &velocity1x,
-                                                 int &velocity1y,
-                                                 int &velocity1z,
-                                                 int &acceleration1x,
-                                                 int &acceleration1y,
-                                                 int &acceleration1z,
-                                                 int &position2x,
-                                                 int &position2y,
-                                                 int &position2z,
-                                                 int &velocity2x,
-                                                 int &velocity2y,
-                                                 int &velocity2z,
-                                                 int &acceleration2x,
-                                                 int &acceleration2y,
-                                                 int &acceleration2z)> &algorithm);
+    void set_algorithm (const std::function<bool(float &position1x,
+                                                 float &position1y,
+                                                 float &position1z,
+                                                 float &velocity1x,
+                                                 float &velocity1y,
+                                                 float &velocity1z,
+                                                 float &acceleration1x,
+                                                 float &acceleration1y,
+                                                 float &acceleration1z,
+                                                 float &position2x,
+                                                 float &position2y,
+                                                 float &position2z,
+                                                 float &velocity2x,
+                                                 float &velocity2y,
+                                                 float &velocity2z,
+                                                 float &acceleration2x,
+                                                 float &acceleration2y,
+                                                 float &acceleration2z)> &algorithm);
     virtual unsigned long get_particle_count ()                           = 0;
     virtual void serialize (std::shared_ptr<ParticleFileWriter> p_writer) = 0;
 };
